requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.1.0
gunicorn==23.0.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
